+++ packages/orq-rc/src/orq_ai_sdk/models/patch_v2_human_eval_sets_id_op.py
@@ -228,7 +228,7 @@ class PatchV2HumanEvalSetsIDResponseBody2(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-13T12:08:27.769Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-13T13:53:07.391Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
@@ -313,7 +313,7 @@ class PatchV2HumanEvalSetsIDResponseBody1(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-13T12:08:27.769Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-13T13:53:07.391Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")