+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRGKTFX9KMVXPJ9YB51ZK6VT"
+        "01KRGST59FY4ETE90CV2KFDYK9"
     )
     r"""The unique identifier of the data source"""