+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -1808,7 +1808,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KRGKTFD2JK7M2Y1CVAETFRSN"
+    id: Optional[str] = "01KRGST4BZHXS5WERASGGDDR0E"
 
     description: Optional[str] = None