+++ packages/orq-rc/RELEASES.md
@@ -7889,3 +7889,13 @@ Based on:
 - [python v4.10.0-rc.2] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.2 - packages/orq-rc
+
+## 2026-05-14 04:59:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.0 (2.884.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.3] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.3 - packages/orq-rc